:zap: att

+++ src/controllers/userController.ts
@@ -1,36 +1,50 @@
+// src/controllers/userController.ts
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import User from '@/models/userModel';
 import connectToDatabase from '@/config/database';
 import { NextApiRequest, NextApiResponse } from 'next';
 
-export const getUsers = async (req: NextApiRequest, res: NextApiResponse) => {
+// get all users
+export const getAllUsers = async (req: NextApiRequest, res: NextApiResponse) => {
+  // Conectar ao banco de dados
   await connectToDatabase();
   try {
+    // Encontrar todos os usuários no banco de dados
     const users = await User.find({});
+    // Retornar os usuários encontrados como resposta
     res.status(200).json(users);
   } catch (e) {
+    // Se ocorrer algum erro, enviar uma resposta de erro com status 500
     res.status(500).json({ message: 'Erro no controller ao buscar todos os usuarios' })
   }
 };
 
 // login
 export const loginUser = async (req: NextApiRequest, res: NextApiResponse) => {
-
-  const secretKey = process.env.JWT_TOKEN;
-
-  if (!secretKey) {
-    return res.status(500).json({ message: 'Chave JWT não definida' });
-  }
   
+  // Chave secreta para assinar o token JWT
+  const JWT_SECRET = process.env.JWT_SECRET;
+
+  // Verificar se JWT_SECRET está definido
+  if (!JWT_SECRET) {
+    throw new Error('JWT_SECRET não está definido.');
+  }
+
+  // verifica se o método é POST
   if (req.method !== 'POST') {
     return res.status(405).json({ message: 'Método não permitido' });
   }
 
+  const { email, password } = req.body;
+
+  // Verificar se o email e a senha foram fornecidos
+  if (!email || !password) {
+    return res.status(400).json({ message: 'Por favor, forneça email e senha.' });
+  }
+  
   // Conecta ao banco de dados
   await connectToDatabase();
-
-  const { email, password } = req.body;
 
   try {
     // Encontra o usuário pelo email
@@ -38,17 +52,17 @@ export const loginUser = async (req: NextApiRequest, res: NextApiResponse) => {
 
     // Verifica se o usuário existe
     if (!user) {
-      return res.status(401).json({ message: 'Credenciais inválidas' });
+      return res.status(401).json({ message: 'Usuário não encontrado.' });
     }
 
     // Verifica se a senha é válida
     const isPasswordValid = await bcrypt.compare(password, user.password);
     if (!isPasswordValid) {
-      return res.status(401).json({ message: 'Credenciais inválidas' });
+      return res.status(401).json({ message: 'Senha inválida' });
     }
 
-    // Gera o token JWT
-    const token = jwt.sign({ userId: user._id }, secretKey, { expiresIn: '1h' });
+    // Gerar um token JWT com o ID do usuário como payload
+    const token = jwt.sign({ userId: user._id }, JWT_SECRET, { expiresIn: '1h' });
 
     // Retorna o token JWT
     return res.status(200).json({ token });
@@ -59,32 +73,31 @@ export const loginUser = async (req: NextApiRequest, res: NextApiResponse) => {
 }
 
 // create user
-export const register = async (req: NextApiRequest, res: NextApiResponse) => {
+export const registerUser = async (req: NextApiRequest, res: NextApiResponse) => {
   // Conecta ao banco de dados
   await connectToDatabase();
-
+  // Extrair dados do corpo da requisição
   const { email, password } = req.body;
-
   try {
-    // Verifica se o usuário já existe no banco de dados
+    // Verifica se o usuário já existe no banco de dados; Verificar se o email já está em uso
     const existingUser = await User.findOne({ email });
     if (existingUser) {
+      // Se o email já estiver em uso, enviar uma resposta com status 400 (Bad Request)
       return res.status(400).json({ message: 'O usuário já existe' });
     }
-
     // Gera um hash da senha
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-
+    const hashedPassword = await bcrypt.hash(password, 10);
     // Cria uma nova instância do usuário com a senha hashada
     const newUser = new User({ email, password: hashedPassword });
-
     // Salva o novo usuário no banco de dados
     await newUser.save();
-
     // Retorna uma resposta de sucesso
-    return res.status(201).json({ message: 'Usuário cadastrado com sucesso' });
+    res.status(201).json({ 
+      message: 'Usuário cadastrado com sucesso.', 
+      user: newUser 
+    });
   } catch (error) {
+    // Se ocorrer algum erro, enviar uma resposta de erro com status 500
     console.error('Erro no cadastro de usuário:', error);
     return res.status(500).json({ message: 'Erro no servidor' });
   }
@@ -94,19 +107,19 @@ export const register = async (req: NextApiRequest, res: NextApiResponse) => {
 export const deleteUser = async (req: NextApiRequest, res: NextApiResponse) => {
   // Conecta ao banco de dados
   await connectToDatabase();
-
+  // const { id } = req.query;
   const userId = req.query.id as string; // Obtém o ID do usuário a ser excluído da consulta
-
+  if (!userId) {
+    return res.status(400).json({ message: 'ID do usuário não fornecido.' });
+  }
   try {
     // Verifica se o usuário existe no banco de dados
     const existingUser = await User.findById(userId);
     if (!existingUser) {
       return res.status(404).json({ message: 'Usuário não encontrado' });
     }
-
     // Exclui o usuário do banco de dados
     await User.findByIdAndDelete(userId);
-
     // Retorna uma resposta de sucesso
     return res.status(200).json({ message: 'Usuário excluído com sucesso' });
   } catch (error) {
